restore int + checks

+++ token/core/zkatdlog/crypto/rp/bulletproof_test.go
@@ -22,8 +22,8 @@ var _ = Describe("Range Proof", func() {
 		Context("If IPA is generated correctly", func() {
 			It("Succeeds", func() {
 				curve := math.Curves[1]
-				nr := uint(3)
-				l := uint(math2.Pow(2, float64(nr)))
+				nr := 3
+				l := int(math2.Pow(2, float64(nr)))
 				leftGens := make([]*math.G1, l)
 				rightGens := make([]*math.G1, l)
